moved buy buy procedure to aquarium, changed coin movement to double, changed fish hunger to double, update some main function

+++ Coin.h
@@ -31,7 +31,7 @@ class Coin : public Object, Moveable {
 		
 		void draw();
 		void remove();
-		void move(int);
+		void move(double);
 	
 	private:
 		int coin_id;
